requirements.txt
anyio==4.0.0
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.1.0
cycler==0.11.0
Django==4.2.4
exceptiongroup==1.1.3
fonttools==4.41.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
image==1.5.33
imageio==2.31.1
importlib-resources==6.0.0
imutils==0.5.4
install==1.3.5
joblib==1.3.1
kiwisolver==1.4.4
lazy_loader==0.3
matplotlib==3.7.2
networkx==3.1
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
PassportEye==2.2.1
pdfminer==20191125
Pillow==10.0.0
prettytable==3.9.0
pycryptodome==3.18.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-telegram-bot==20.4
PyWavelets==1.4.1
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
sqlean.py==0.21.8.4
sqlite-sqlean @ git+https://github.com/karlb/sqlite-sqlean@0b99f9c61f57740346a86c6148f5c1d7d31dac2c
sqlparse==0.4.4
tesseract==0.1.3
threadpoolctl==3.2.0
tifffile==2023.7.18
typing_extensions==4.7.1
urllib3==2.1.0
wcwidth==0.2.7
zipp==3.16.2
